agile_be/internship-main/requirements.txt
Django
wagtail
djangorestframework 
djangorestframework-simplejwt
django-cors-headers
mysqlclient  
reportlab>=3.6.12
pytz==2025.2